Add files via upload
With this script you should end with a csv file ""Preprocessed_train.csv", in which the feature "orig_destination_distance" was deleted, the NaN in the features prop_review_score and prop_location_score2 replace by zeros. The NaN in feature srch_query_affinity_score replaced by the worst case scenario, in this case the minimum value -326.567500.